fix: add comments and rename camera param to camera param map
