+++ Camera/CamerasPhantom/AmetekS710EuresysCoaxlinkOcto/CMakeLists.txt
@@ -4,7 +4,7 @@ add_library("${PROJECT_NAME}" SHARED
     camera_phantom_s710.cc
     ../../Camera/camera_logger.cc
     ../camera_phantom_interface.cc
-    ../camera_param.cc
+    ../camera_param_map.cc
 )
 
 target_include_directories("${PROJECT_NAME}" PRIVATE
